fix(prd-admin): 设置页名称与侧栏统一 + 修复加分隔无效
1. 抽取 getShortLabel + SHORT_LABEL_MAP 到 lib/shortLabel.ts，
   AppShell 与设置页「我的导航/候选池」芯片共用同一份短标签规则。
   用户在侧栏看到"百宝箱"，设置页芯片也显示"百宝箱"，不再出现
   "AI 百宝箱"这种全名/简称对不上的错位。

2. 修复"加分隔"按钮点击无反应：原逻辑将分隔符追加到数组末尾，
   但 collapseDividers 会剥掉真正的尾部分隔符（尾部分隔符无意义），
   结果 setNavLayout 下去的还是原数组。
   改为在最后一个条目之前插入分隔符，用户点击后即可看到新横杆
   出现在倒数第二位，随后可拖动到任意位置。
