+++ prd-admin/src/layouts/AppShell.tsx
@@ -50,6 +50,7 @@ import { useThemeStore } from '@/stores/themeStore';
 import { useLayoutStore } from '@/stores/layoutStore';
 import { useNavOrderStore } from '@/stores/navOrderStore';
 import { getLauncherCatalog } from '@/lib/launcherCatalog';
+import { getShortLabel } from '@/lib/shortLabel';
 import { useBreakpoint } from '@/hooks/useBreakpoint';
 import { SystemDialogHost } from '@/components/ui/SystemDialogHost';
 import { InlinePageLoader } from '@/components/ui/VideoLoader';
@@ -75,52 +76,8 @@ const NAV_GROUPS: { key: string; label: string }[] = [
   { key: 'admin', label: '系统管理' },
 ];
 
-/** 折叠态短标签映射（2-4 字） */
-const SHORT_LABEL_MAP: Record<string, string> = {
-  'ai-toolbox': '百宝箱',
-  'report-agent': '周报',
-  'workflow-agent': '工作流',
-  'marketplace': '市场',
-  'my-resources': '我的资源',
-  'my-assets': '我的资源',
-  'model-center': '模型',
-  'mds': '模型',
-  'authz': '用户权限',
-  'users': '用户',
-  'data-ops': '自定义',
-  'settings': '自定义',
-  'visual-agent': '视觉',
-  'literary-agent': '文学',
-  'video-agent': '视频',
-  'defect-agent': '缺陷',
-  'prd-agent': '智能体',
-  'arena-agent': '竞技场',
-  'shortcuts-agent': '快捷指令',
-  'data-migration-agent': '迁移',
-  'executive': '团队',
-  'tutorial-email': '邮件',
-  'lab': '实验室',
-  'automations': '自动化',
-  'skills': '技能',
-  'dashboard': '仪表盘',
-  'groups': '群组',
-  'prompts': '提示词',
-  'assets': '资源',
-  'logs': '日志',
-  'data': '数据',
-  'open-platform': '开放平台',
-};
-
 /** 从侧边栏隐藏的 appKey（页面仍可直接访问） */
 const HIDDEN_NAV_KEYS = new Set<string>([]);
-
-/** 获取短标签：优先查映射表，否则使用完整 label */
-function getShortLabel(appKey: string, label: string): string {
-  if (SHORT_LABEL_MAP[appKey]) return SHORT_LABEL_MAP[appKey];
-  // 去掉常见后缀
-  const clean = label.replace(/\s*(Agent|管理|引擎)\s*/g, '').trim();
-  return clean.length <= 4 ? clean : clean.slice(0, 4);
-}
 
 /** 根据 mimeType 推断扩展名，确保下载文件名带后缀 */
 function ensureDownloadName(name: string | undefined | null, mimeType?: string | null): string {
